Allow repotool tests to fail gracefully when proper VCS is not installed.

+++ test/repotool-checkout-git.sh
@@ -1,5 +1,7 @@
 ## Test repotool checkout of git repo
 
+command -v git >/dev/null 2>&1 || { echo "    Skipped, git missing."; exit 0; }
+
 trap "rm -rf /tmp/test-repo$$ /tmp/target$$ /tmp/out$$" 0 12 2 15
 
 ./fi-to-fi -n /tmp/test-repo$$ < simple.fi
@@ -13,7 +15,7 @@ case $1 in
     --regress)
         diff --text -u $2.chk /tmp/out$$ || exit 1; ;;
     --rebuild)
-	cat /tmp/out$$ ;;
+	cat /tmp/out$$ >$2.chk;;
 esac
 	      
 #end
