+++ test/repotool-compare-tags-git-hg.sh
@@ -1,6 +1,9 @@
 ## Test comparing tags between git and Mercurial repository
 # Reproduces https://gitlab.com/esr/reposurgeon/issues/39
 
+command -v git >/dev/null 2>&1 || { echo "    Skipped, git missing."; exit 0; }
+command -v hg >/dev/null 2>&1 || { echo "    Skipped, hg missing."; exit 0; }
+
 trap "rm -rf /tmp/test-repo$$-git /tmp/test-repo$$-hg /tmp/out$$" 0 12 2 15
 
 ./fi-to-fi -n /tmp/test-repo$$-git < lighttag.fi
@@ -11,6 +14,6 @@ case $1 in
     --regress)
         diff --text -u $2.chk /tmp/out$$ || exit 1; ;;
     --rebuild)
-	cat /tmp/out$$ ;;
+	cat /tmp/out$$ >$2.chk;;
 esac
