+++ test/repotool-tags-hg.sh
@@ -1,5 +1,7 @@
 ## Test listing tags in a Mercurial repository
 
+command -v hg >/dev/null 2>&1 || { echo "    Skipped, hg missing."; exit 0; }
+
 trap "rm -rf /tmp/test-repo$$ /tmp/target$$ /tmp/out$$" 0 12 2 15
 
 ./hg-to-fi -n /tmp/test-repo$$ < lighttag.fi
@@ -10,7 +12,7 @@ case $1 in
     --regress)
         diff --text -u $2.chk /tmp/out$$ || exit 1; ;;
     --rebuild)
-	cat /tmp/out$$ ;;
+	cat /tmp/out$$ >$2.chk;;
 esac
 	      
 #end
